requirements.txt
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
